Консольное Приложение №1 Исправлен выход на бесконечность в расчете разделеного счета Разнесен код по разным классам
